Now, KeepAlive building block is an actor.
